+++ src/openrct2/scenario/Scenario.cpp
@@ -68,7 +68,7 @@ bool gFirstTimeSaving = true;
 uint16_t gSavedAge;
 uint32_t gLastAutoSaveUpdate = 0;
 
-uint64_t gScenarioTicks;
+uint32_t gScenarioTicks;
 random_engine_t gScenarioRand;
 
 Objective gScenarioObjective;
